a2afinaldemo/pyproject.toml
[project]
name = "a2a_multi_agent"
version = "0.1.0"
description = "Agent-to-agent demo using LangChain, CrewAI, AutoGen, and A2A SDK with Streamlit UI"
authors = [{name = "Prashil"}]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.11",
    "autogen-agentchat>=0.7.5",
    "crewai>=1.4.1",
    "fastapi>=0.115.0",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.2",
    "matplotlib>=3.10.0",
    "pandas>=2.3.0",
    "plotly>=6.4.0",
    "python-dotenv>=1.0.1",
    "pyautogen>=0.10.0",
    "sse-starlette>=2.1.0",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
]

[tool.uv]
package = true

